refactor(shell): extract wingman panel shell
Split panel DOM refs, tab switching, open/toggle/layout, and resize
handling from wingman/index.js into wingman/panel.js. Panel module
exposes an initPanel({ renderer, hooks }) factory so the handoff layer
in index.js can still influence preferred-tab-on-open and react to
panel open/close events without importing handoff state into panel.js.

window exports unchanged: openWingmanPanel, toggleWingmanPanel,
updatePanelLayout — now set inside panel.js (mirroring how alerts.js
owns dismissAlert).
